signals

+++ owner/models.py
@@ -61,13 +61,13 @@ class Owner (AbstractBaseUser):
 
 
 # signals 
-
 @receiver(pre_save, sender = get_user_model())
 def user_check(instance , sender, signal, **kwArgs ):
-    email = instance.email
-    user=get_user_model().objects.filter(email=email,confirmed=True).first()
-    if user:
-        user.delete()
-        print(user,'user slresdy exists and as delted ')
-    else:
-        print('user created without anything ')
+    if not instance.pk:
+        email = instance.email
+        user=get_user_model().objects.get(email=email,confirmed=True)
+        if user:
+            user.delete()
+            print(user,'user slresdy exists and as delted ')
+        else:
+            print('user created without anything ')
